+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-...
+### Added 
+
+- Oracle support for Basic and Calendar table aggregates
 
 ## [0.9.2] - 2019-07-04